+++ prowler/providers/aws/services/mq/mq_broker_cluster_deployment_mode/mq_broker_cluster_deployment_mode.metadata.json
@@ -1,6 +1,6 @@
 {
   "Provider": "aws",
-  "CheckID": "mq_rabbitmq_broker_cluster_mode",
+  "CheckID": "mq_broker_cluster_deployment_mode",
   "CheckTitle": "MQ RabbitMQ Brokers should use cluster deployment mode.",
   "CheckType": [
     "Software and Configuration Checks/Industry and Regulatory Standards/NIST 800-53 Controls"
@@ -12,17 +12,17 @@
   "ResourceType": "AwsAmazonMQBroker",
   "Description": "Ensure that RabbitMQ Brokers use cluster deployment mode.",
   "Risk": "Using a single-instance RabbitMQ broker limits fault tolerance and high availability. Without cluster deployment, broker failures could lead to significant downtime and potential data loss.",
-  "RelatedUrl": "https://docs.aws.amazon.com/amazon-mq/latest/developer-guide/amazon-mq-broker-architecture.html",
+  "RelatedUrl": "https://docs.aws.amazon.com/amazon-mq/latest/developer-guide/rabbitmq-basic-elements.html",
   "Remediation": {
     "Code": {
-      "CLI": "aws mq create-broker --broker-name <your-broker-name> --engine-type RabbitMQ --deployment-mode CLUSTER_MULTI_AZ --auto-minor-version-upgrade --engine-version <version> --host-instance-type <instance-type> --publicly-accessible --subnet-ids <subnet-ids>",
+      "CLI": "aws mq create-broker --broker-name <your-broker-name> --engine-type RabbitMQ --deployment-mode CLUSTER_MULTI_AZ",
       "NativeIaC": "",
       "Other": "https://docs.aws.amazon.com/securityhub/latest/userguide/mq-controls.html#mq-6",
       "Terraform": ""
     },
     "Recommendation": {
       "Text": "Ensure RabbitMQ brokers are deployed in cluster mode to enhance resilience and prevent data loss during failures.",
-      "Url": "https://docs.aws.amazon.com/amazon-mq/latest/developer-guide/getting-started-rabbitmq.html"
+      "Url": "https://docs.aws.amazon.com/amazon-mq/latest/developer-guide/rabbitmq-broker-architecture.html#rabbitmq-broker-architecture-cluster"
     }
   },
   "Categories": [],
